requirements.txt
gensim==4.3.3
nltk==3.9.1
numpy>=1.18.5,<2.0,
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.2
setuptools==75.1.0
